+++ src/app/app.routes.ts
@@ -1,18 +1,21 @@
 import { Routes } from '@angular/router';
-import { ProjectsComponent } from './pages/projects/projects.component';
-import { FacsimilesComponent } from './pages/facsimiles/facsimiles.component';
-import { PersonsComponent } from './pages/persons/persons.component';
-import { LoginComponent } from './pages/login/login.component';
+
 import { authGuard } from './guards/auth.guard';
-import { HomeComponent } from './pages/home/home.component';
-import { PublicationCollectionsComponent } from './pages/publication-collections/publication-collections.component';
-import { NewPublicationFacsimileComponent } from './pages/new-publication-facsimile/new-publication-facsimile.component';
-import { FacsimileCollectionComponent } from './pages/facsimile-collection/facsimile-collection.component';
-import { PublicationBundleComponent } from './pages/publication-bundle/publication-bundle.component';
 import { AddFacsimileFromPublicationsComponent } from './pages/add-facsimile-from-publications/add-facsimile-from-publications.component';
-import { TableOfContentsComponent } from './pages/table-of-contents/table-of-contents.component';
+import { FacsimileCollectionComponent } from './pages/facsimile-collection/facsimile-collection.component';
+import { FacsimileCollectionUploadBlockComponent } from './pages/facsimile-collection-upload-block/facsimile-collection-upload-block.component';
+import { FacsimileCollectionUploadSelectionComponent } from './pages/facsimile-collection-upload-selection/facsimile-collection-upload-selection.component';
+import { FacsimilesComponent } from './pages/facsimiles/facsimiles.component';
+import { HomeComponent } from './pages/home/home.component';
 import { KeywordsComponent } from './pages/keywords/keywords.component';
 import { KeywordLinkingComponent } from './pages/keyword-linking/keyword-linking.component';
+import { LoginComponent } from './pages/login/login.component';
+import { NewPublicationFacsimileComponent } from './pages/new-publication-facsimile/new-publication-facsimile.component';
+import { PersonsComponent } from './pages/persons/persons.component';
+import { ProjectsComponent } from './pages/projects/projects.component';
+import { PublicationBundleComponent } from './pages/publication-bundle/publication-bundle.component';
+import { PublicationCollectionsComponent } from './pages/publication-collections/publication-collections.component';
+import { TableOfContentsComponent } from './pages/table-of-contents/table-of-contents.component';
 
 export const routes: Routes = [
   { path: '', component: HomeComponent, canActivate: [authGuard] },
@@ -25,6 +28,9 @@ export const routes: Routes = [
   { path: 'publication-collections/:collectionId/publication/:publicationId/facsimiles/new', component: NewPublicationFacsimileComponent, canActivate: [authGuard] },
   { path: 'facsimiles', component: FacsimilesComponent, canActivate: [authGuard] },
   { path: 'facsimiles/add-from-publications', component: AddFacsimileFromPublicationsComponent, canActivate: [authGuard] },
+  { path: 'facsimiles/:id/upload-selection', component: FacsimileCollectionUploadSelectionComponent, canActivate: [authGuard] },
+  { path: 'facsimiles/:id/upload-missing', component: FacsimileCollectionUploadBlockComponent, canActivate: [authGuard], data: { mode: 'missing' } },
+  { path: 'facsimiles/:id/upload-all', component: FacsimileCollectionUploadBlockComponent, canActivate: [authGuard], data: { mode: 'all' } },
   { path: 'facsimiles/:id', component: FacsimileCollectionComponent, canActivate: [authGuard] },
   { path: 'keywords', component: KeywordsComponent, canActivate: [authGuard] },
   { path: 'keywords/linking', component: KeywordLinkingComponent, canActivate: [authGuard] },