+++ src/app/pages/facsimile-collection/facsimile-collection.component.ts
@@ -1,57 +1,73 @@
 import { CommonModule } from '@angular/common';
 import { HttpErrorResponse } from '@angular/common/http';
-import { Component, OnInit } from '@angular/core';
+import { Component, computed, inject, OnInit, signal } from '@angular/core';
 import { ActivatedRoute, RouterLink } from '@angular/router';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
-import { Observable, take } from 'rxjs';
+import { Observable, finalize, take, tap } from 'rxjs';
 
-import { FileUploadComponent } from '../../components/file-upload/file-upload.component';
 import { FacsimileCollection, VerifyFacsimileFileResponse } from '../../models/facsimile.model';
 import { FacsimileService } from '../../services/facsimile.service';
 import { ProjectService } from '../../services/project.service';
-import { RangeArrayPipe } from '../../pipes/range-array.pipe';
+import { LoadingSpinnerComponent } from '../../components/loading-spinner/loading-spinner.component';
 
 @Component({
   selector: 'facsimile-collection',
   imports: [
-    MatIconModule, FileUploadComponent, MatButtonModule, RouterLink, CommonModule,
-    RangeArrayPipe
+    CommonModule,
+    RouterLink,
+    MatButtonModule,
+    MatIconModule,
+    LoadingSpinnerComponent
   ],
   templateUrl: './facsimile-collection.component.html',
   styleUrl: './facsimile-collection.component.scss'
 })
 export class FacsimileCollectionComponent implements OnInit {
-  collectionId: number;
-  facsimile$: Observable<FacsimileCollection> = new Observable<FacsimileCollection>();
-  missingFileNumbers: number[] = [];
+  private readonly facsimileService = inject(FacsimileService);
+  private readonly projectService = inject(ProjectService);
+  private readonly route = inject(ActivatedRoute);
 
-  constructor(
-    private fascimileService: FacsimileService, 
-    private projectService: ProjectService,
-    private route: ActivatedRoute
-  ) {
-    this.collectionId = this.route.snapshot.params['id'];
-  }
+  collectionId: number = this.route.snapshot.params['id'];
+  facsimile$: Observable<FacsimileCollection> = new Observable<FacsimileCollection>();
+  loadingFacsData = signal<boolean>(true);
+  missingFileNumbers = signal<number[]>([]);
+  missingFileSet = computed(() => new Set(this.missingFileNumbers()));
+  numberOfPages = signal<number>(0);
+  pageNumbers = computed(() =>
+    Array.from({ length: this.numberOfPages() }, (_, i) => i + 1)
+  );
+  project: string | null = this.projectService.getCurrentProject();
 
   ngOnInit() {
-    const currentProject = this.projectService.getCurrentProject();
-    this.facsimile$ = this.fascimileService.getFacsimileCollection(this.collectionId, currentProject);
+    this.facsimile$ = this.facsimileService.getFacsimileCollection(
+      this.collectionId,
+      this.project
+    ).pipe(
+      tap((facsColl) => {
+        this.numberOfPages.set(facsColl.number_of_pages ?? 0);
+      }),
+      finalize(() => this.loadingFacsData.set(false))
+    );
     this.verifyFacsimileFiles();
   }
 
   verifyFacsimileFiles() {
-    const currentProject = this.projectService.getCurrentProject();
-    this.fascimileService.verifyFacsimileFile(this.collectionId, 'all', currentProject).pipe(
+    this.facsimileService.verifyFacsimileFile(
+      this.collectionId,
+      'all',
+      this.project
+    ).pipe(
       take(1)
     ).subscribe({
       next: response => {
-        this.missingFileNumbers = response.data?.missing_file_numbers || [];
+        this.missingFileNumbers.set(response.data?.missing_file_numbers || []);
       },
       error: (error: HttpErrorResponse) => {
         const err: VerifyFacsimileFileResponse = error.error;
-        this.missingFileNumbers = err.data?.missing_file_numbers || [];
+        this.missingFileNumbers.set(err.data?.missing_file_numbers || []);
       }
     });
   }
+
 }